src/requirements.txt
torch==1.12.1
torchvision==0.13.1
torchaudio==0.12.1
fastai==2.7.9
ipykernel==6.16.0
opencv-python==4.6.0.66
python-dotenv==0.21.0